Fix GitHub Action when run in other repos
The underlying problem was that the current working directory has to be within the github action's directory in order to pick up yarn v2.
This happens automatically in our own test job because the default cwd is the repo's directory which _is_ our github action's directory.
But in another user's repo, the cwd is the user's repo root. So when we spawn 'yarn node' we have to set cwd to our own github action directory.